+++ web3/src/main/java/org/hiero/mirror/web3/state/Utils.java
@@ -2,14 +2,14 @@
 
 package org.hiero.mirror.web3.state;
 
+import static org.hiero.mirror.common.util.DomainUtils.isLongZeroAddress;
+
 import com.hedera.hapi.node.base.Key;
 import com.hedera.hapi.node.base.KeyList;
 import com.hedera.hapi.node.base.Timestamp;
 import com.hedera.pbj.runtime.ParseException;
 import com.hedera.pbj.runtime.io.buffer.Bytes;
-import java.nio.ByteBuffer;
 import java.time.Instant;
-import java.util.Arrays;
 import lombok.experimental.UtilityClass;
 import org.hyperledger.besu.datatypes.Address;
 
@@ -44,21 +44,7 @@ public class Utils {
         return new Timestamp(instant.getEpochSecond(), instant.getNano());
     }
 
-    public boolean isMirror(final Address address) {
-        return address != null && isMirror(address.toArrayUnsafe());
-    }
-
-    public static boolean isMirror(final byte[] address) {
-        if (address == null || address.length != EVM_ADDRESS_LEN) {
-            return false;
-        }
-
-        if (Arrays.equals(address, Address.ZERO.toArrayUnsafe())) {
-            return true;
-        }
-
-        var wrapper = ByteBuffer.wrap(address);
-
-        return wrapper.getInt() == 0 && wrapper.getLong() == 0;
+    public static boolean isMirror(final Address address) {
+        return address != null && isLongZeroAddress(address.toArrayUnsafe());
     }
 }
